WP-CLI update from 0.23.1 to 0.24+
The reason we had WP-CLI package locked to a particular reason
was due to its compatibility issues with the WordPress itself.

Version 0.24.0 removes the dependency on the wp-settings.php of
WordPress and therefor it can be used with any version of
WordPress.  Removing a particular version lock should simplify
the maintenance and upgrades in the future.

For more information, read the following:
http://wp-cli.org/blog/version-0.24.0.html