Added robots.php script and ALLOW_ROBOTS support (task #2696)
Added `robots.php` script which allows for dynamic robots.txt,
based on the ALLOW_ROBOTS setting in the `.env`.

If the ALLOW_ROBOTS is not defined or set to false, the response
will deny indexing of anything on the site.  If it is set to true
and the `robots.txt` file exists, then those rules will be used.
If it is set to true and `robots.txt` does not exist or is not
readable, then a default response will allow indexing of anything
on the site.